feat: include cmdb plugin
Include the cmdb plugin in the image.

By including the cmdb plugin, it makes continued testing of its use
easier given the auto-detection logic in the bash executor.

Note that because the cmdb plugin is not currently included in the
`GENERATION_PLUGIN_DIRS` environment variable, dynamic cmdb processing
won't be active by default, but simply by adding it as an extra path
value to this variable, dynamic cmdb loading will be activated.

Also add the diagrams plugin to align with what is in the docker
container build.